feat: rewrite sd-accordion-group documentation
